requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlmodel>=0.0.16
python-dotenv>=1.0
pydantic>=2.6
requests>=2.31

streamlit>=1.31
plotly>=5.18
pandas>=2.1
